requirements.txt
google-adk==1.3.0
google-cloud-aiplatform[agent_engines,adk]==1.97.0
cloudpickle==3.1.1
pydantic==2.11.6
PyMuPDF==1.26.1
python-docx==1.1.2
python-dotenv==1.1.0
requests
